Improve command check
